WIP: Add Scraper/ScraperRun classes
First stab at factoring out a Scraper class and a ScraperRun class.

The IndexToMemebers class is probably overkill for now — we can find the
higher level abstractions like that after we nail down the basics — but
it's probably useful to include in this intial WIP which is more for
feedback than expected to actually merge yet.