Replace IndexToMembers class with RomanianParliament class
Rather than trying to find the correct generic abstraction this forces
the scrapers to have a class that encapsulates navigating a website to
get the correct content.

In the future we can potentially wrap these up in `IndexToMembers` type
classes, but for now it seems easier to just do scraper specific ones
until we find the correct abstraction.